Move some offline cache utils from analysis/ to util/
