+++ taichi/util/CMakeLists.txt
@@ -9,6 +9,7 @@ target_sources(taichi_util
     image_buffer.cpp
     image_io.cpp
     lang_util.cpp
+    offline_cache.cpp
     short_name.cpp
     statistics.cpp
     str.cpp
