+++ taichi/util/offline_cache.h
@@ -10,6 +10,7 @@
 #include "taichi/common/core.h"
 #include "taichi/common/cleanup.h"
 #include "taichi/common/version.h"
+#include "taichi/rhi/arch.h"
 #include "taichi/util/io.h"
 #include "taichi/util/lock.h"
 
@@ -285,6 +286,13 @@ class CacheCleaner {
     }
   }
 };
+
+std::string get_cache_path_by_arch(const std::string &base_path, Arch arch);
+bool enabled_wip_offline_cache(bool enable_hint);
+std::string mangle_name(const std::string &primal_name, const std::string &key);
+bool try_demangle_name(const std::string &mangled_name,
+                       std::string &primal_name,
+                       std::string &key);
 
 }  // namespace offline_cache
 }  // namespace lang
